refactor: split CAPPI into dedicated module
