+++ cinrad/__init__.py
@@ -5,7 +5,7 @@ from . import calc
 from . import visualize
 from . import correct
 
-from .calc import CAPPI
+from .cappi import CAPPI
 from .io import read_level2
 
 __version__ = "1.9.3"
